pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "service-catalog-sync"
version = "0.1.0"
description = "Automated Backstage catalog hydration — scans repos, infers ownership and tech stack, keeps catalog fresh via CI"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
authors = [{ name = "Brian Pelow", email = "brianpelow@github" }]
keywords = ["backstage", "catalog", "platform-engineering", "innersource", "fintech"]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "pyyaml>=6.0.1",
    "gitpython>=3.1.43",
]

[project.scripts]
catalog-sync = "catalogsync.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/catalogsync"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]